feat: create UI for auth link
