+++ next/src/app/components/organisms/layout/Header.tsx
@@ -1,6 +1,6 @@
 'use client'
 
-import { HStack, Spacer } from '@chakra-ui/react'
+import { Box, Button, HStack, Spacer } from '@chakra-ui/react'
 import Logo from '/public/images/logo.svg'
 import { SignedIn, SignedOut, UserButton } from '@clerk/nextjs'
 import NextLink from 'next/link'
@@ -8,23 +8,28 @@ import CustomIcon from '../../atoms/CustomIcon'
 
 const Header = () => {
   return (
-    <header>
+    <Box as='header'>
       <HStack p={4} bg="brand.secondary">
-        <NextLink href="#">
+        <NextLink href="/">
           <Logo />
         </NextLink>
         <Spacer />
-        <SignedIn>
-          <UserButton />
-        </SignedIn>
-        <SignedOut>
-          <NextLink href="/sign-in">ログイン</NextLink>
-        </SignedOut>
-        <NextLink href="#">
-          <CustomIcon iconName="FaBookmark" color="brand.brown" />
-        </NextLink>
+        <Box>
+          <SignedIn>
+            <UserButton />
+            <NextLink href="#">
+              <CustomIcon iconName="FaBookmark" color="brand.brown" />
+            </NextLink>
+          </SignedIn>
+          <SignedOut>
+            <Button size='sm' variant="outline" as={NextLink} href="./sign-in" gap='2px'>
+              ログイン
+              <CustomIcon iconName="FaSignInAlt" />
+            </Button>
+          </SignedOut>
+        </Box>
       </HStack>
-    </header>
+    </Box>
   )
 }
