+++ next/src/app/theme/components/button.ts
@@ -7,5 +7,15 @@ export const Button = {
     secondary: {
       bg: 'brand.gray',
     },
+    outline: {
+      bg: 'white',
+      borderColor: 'brand.primary',
+      borderWidth: 2,
+      color: 'brand.primary',
+      _hover: {
+        bg: 'brand.primary',
+        color: 'white',
+      }
+    }
   },
 }
